requirements.txt
Flask==3.1.0
ipython==8.31.0
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.37
